fix(dotcom): dismiss-only clicks on the editor top panels while a menu is open
Clicking the editor's top panels (the top-left menu / file name, the
top-right share / people buttons) while any menu was open both dismissed
the menu and acted on what was pressed — e.g. opening the share menu.
The SDK's MenuClickCapture only covers the canvas (z 250, below the
panels at 300) and the sidebar had its own overlay, but the top panels
had none.

Generalize the sidebar's dismiss-only overlay into a shared
TlaMenuClickCapture and render it inside the top panels too. While any
menu is open it sits above the panel's buttons but below the portaled
menu and swallows the press, so the press only dismisses. It never
covers the canvas, so the SDK capture keeps owning click-to-draw and
drag — no SDK changes needed.
